Refactor catalog service with base one
