+++ browser/index.d.ts
@@ -1,2 +1,2 @@
 import WakandaClient from '../dist/wakanda-client';
-export default WakandaClient;
+export {WakandaClient};
