+++ src/entry.node.ts
@@ -3,4 +3,4 @@ import NodeHttpClient from './data-access/http/node-http-client';
 
 WakandaClient.HttpClient = NodeHttpClient;
 
-export = WakandaClient;
+export {WakandaClient};
